refactor job specific classes to be more unit testable clear separation of gql queries to mock.
