+++ fflogs_rotation/dragoon.py
@@ -35,8 +35,6 @@ class DragoonActions(BuffQuery):
         self.life_of_the_dragon_id = life_of_the_dragon_id
         if patch_number < 7.0:
             self._set_combo_finisher_timings(headers)
-        else:
-            self._set_life_of_the_dragon_timings(headers)
 
     def _set_combo_finisher_timings(self, headers: Dict[str, str]) -> None:
         """
@@ -138,54 +136,6 @@ class DragoonActions(BuffQuery):
         self.life_of_the_dragon_times = self.life_of_the_dragon_times.astype(
             int
         ).to_numpy()
-
-    def _set_life_of_the_dragon_timings(self, headers: Dict[str, str]) -> None:
-        """
-        Set the timings for Life of the Dragon based on the provided headers.
-
-        Parameters:
-            headers (Dict[str, str]): Headers for the GraphQL query.
-        """
-        query = """
-        query dragoonLOTD(
-            $code: String!
-            $id: [Int]!
-            $playerID: Int!
-            $lifeOfTheDragonID: Float!
-        ) {
-            reportData {
-                report(code: $code) {
-                    startTime
-                    lifeOfTheDragon: table(
-                            fightIDs: $id
-                            dataType: Buffs
-                            sourceID: $playerID
-                            abilityID: $lifeOfTheDragonID
-                    )
-                }
-            }
-        }
-        """
-
-        variables = {
-            "code": self.report_id,
-            "id": [self.fight_id],
-            "playerID": self.player_id,
-            "lifeOfTheDragonID": self.life_of_the_dragon_id,
-        }
-
-        self._perform_graph_ql_query(headers, query, variables, "dragoonLOTD")
-
-        self.life_of_the_dragon_times = (
-            pd.DataFrame(
-                self.request_response["data"]["reportData"]["report"][
-                    "lifeOfTheDragon"
-                ]["data"]["auras"][0]["bands"]
-            )
-            .astype(int)
-            .to_numpy()
-        )
-        self.life_of_the_dragon_times += self.report_start
 
     def apply_endwalker_combo_finisher_potencies(
         self, actions_df: pd.DataFrame
